requirements.txt
# Core ML/Audio
torch>=1.9.0
torchaudio>=0.10.0
numpy>=1.20.0
scipy>=1.7.0
librosa>=0.8.0
pyworld>=0.3.0
tqdm>=4.60.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration & Logging
PyYAML>=5.4.0
logging>=0.4.9.6

# Optional: Experiment Tracking
# tensorboard>=2.8.0
# wandb>=0.13.0

# Testing
pytest>=6.2.0
pytest-cov>=2.10.0

# Development
black>=21.9b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0